// src/app/dashboard/model/assets-per-day.model.ts
/*
 * Copyright 2021 The PartChain Authors. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 *
 *
 * @export
 * @interface Axis
 */
export interface Axis {
  x: Date;
  y: number;
}

/**
 *
 *
 * @export
 * @interface AssetsPerDay
 */
export interface AssetsPerDay {
  ownAssets: Axis;
  otherAssets: Axis;
}

/**
 *
 *
 * @export
 * @interface Graph
 */
export interface Graph {
  data: Axis[];
  labels: string[];
}

/**
 *
 *
 * @export
 * @interface AssetsPerDayGraph
 */
export interface AssetsPerDayGraph {
  ownAssets: Graph;
  otherAssets: Graph;
}